fixed issue where music slider would bug out when paused and moving. fixed bug where direct downloading a song wouldn't save the title. Todo:
weird performance issue where switching pages is slow on my physical phone, but isn't on the emulator, takes like 2 seconds to switch screens.

add foreground notification controls so you can pause, move around in the song, or eventually skip it from the notification

make playlist functionality in the database that this time only saves the names and playlist they are a part of, and will seperately look into a seperate database with the file information

actually design ui components

if downloaded straight through using the implemented youtube search, save the artists name to the file
figure out how to add artist name anyways

for directDownload method with a url, I need to find a way to find the name of the song, probably the noembed implentation I will add

there is a bug that when player is paused, moving the slider bugs out and resets it to 0 until unpaused.

also need to fix the issue where you can import the same file infinite amount of times

add setting stuff to clear db or things like that