deleted old schema's that were giving me issues. fixed an issue where you couldn't delete songs that have just been added to database
added notification so it turns on once the player is started. and automatically closes when app is fully closed.
Todo:

weird performance issue where switching pages is slow on my physical phone, but isn't on the emulator, takes like 2 seconds to switch screens.

add foreground notification controls so you can pause, move around in the song, or eventually skip it from the notification

make playlist functionality in the database that this time only saves the names and playlist they are a part of, and will seperately look into a seperate database with the file information

actually design ui components

if downloaded straight through using the implemented youtube search, save the artists name to the file
figure out how to add artist name anyways

add setting stuff to clear db or things like that