added playlist full functionality in regards to the music player. when a new screen is loaded in and a song is played, all songs in the playlist are loaded in.
with the full playlist in the player. the notification buttons for skip and previous are fully functional

Todo:

add settings screen with some controls, like clear local database, or colors

when a song is clicked, pull up a song screen where it shows the song, album art, and the full player screen with skip and previous button

figure out how to embed artist name and thumbnail art into id3 tag

actually design ui components

implement playlist downloading, specifically starting with a youtube playlist, and maybe further on a spotify playlist.
spotify will be more complicated because each song needs to be searched for on youtube before downloading